New sprites, backgrounds, instructions, sounds
Added trash, boat, and harpoon sprites for more obstacles. Also added cloud, mountain, sea, and sand backgrounds and tiles from the asset store and have them scroll during gameplay. Added a guide to the title screen with controls and general objectives. Added a sound to be played when the shark is spawned.